Cryptography Project
The aim was to create a multi-user chat based on the client-server model that sends and receives messages. Each user is able to chat to multiple other users simultaneously (many-to-many). This was done in Java using socket programming and telnet as the connection protocol and data was encrypted using the RSA cryptographic algorithm.